fix(PathController): 나이 파라미터 추가
