+++ src/main/java/nextstep/subway/domain/FareCalculatorImpl.java
@@ -1,0 +1,5 @@
+package nextstep.subway.domain;
+
+public interface FareCalculatorImpl {
+    int calculate(int distance);
+}
