+++ src/main/java/nextstep/subway/line/domain/Lines.java
@@ -1,7 +1,5 @@
 package nextstep.subway.line.domain;
 
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 
 public class Lines {
