+++ src/main/java/nextstep/subway/path/ui/PathController.java
@@ -1,5 +1,7 @@
 package nextstep.subway.path.ui;
 
+import nextstep.subway.auth.domain.AuthenticationPrincipal;
+import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.path.application.PathService;
 import nextstep.subway.path.dto.PathResponse;
 import org.springframework.http.ResponseEntity;
@@ -19,8 +21,9 @@ public class PathController {
     }
 
     @GetMapping
-    public ResponseEntity<PathResponse> findShortestPath(@RequestParam(name = "source") Long source,
+    public ResponseEntity<PathResponse> findShortestPath(@AuthenticationPrincipal LoginMember loginMember,
+                                                         @RequestParam(name = "source") Long source,
                                                          @RequestParam(name = "target") Long target) {
-        return ResponseEntity.ok(pathService.findShortestPath(source, target));
+        return ResponseEntity.ok(pathService.findShortestPath(source, target, loginMember.getAge()));
     }
 }
